refactored handlers.go into 5 different files for api data, courses, terms, profs, and grades handlers
