+++ internal/server/handlers/handlers.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/acmutd/acmutd-api/internal/firebase"
-	"github.com/acmutd/acmutd-api/internal/types"
 	"github.com/gin-gonic/gin"
 )
 
@@ -36,197 +35,6 @@ func (h *Handler) Health(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{
 		"status":  "healthy",
 		"message": "ACM API is running",
-	})
-}
-
-// GetAllCourses returns a helpful error since callers must provide a term.
-func (h *Handler) GetAllCourses(c *gin.Context) {
-	c.JSON(http.StatusBadRequest, gin.H{
-		"error": "Term parameter is required. Use /api/v1/courses/{term}",
-	})
-}
-
-// GetCoursesByTerm fetches courses within a term.
-func (h *Handler) GetCoursesByTerm(c *gin.Context) {
-	term := normalizeTerm(c.Param("term"))
-	if term == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term parameter is required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	var (
-		courses []types.Course
-		hasNext bool
-		err     error
-	)
-
-	courses, hasNext, err = h.db.GetAllCoursesByTerm(c.Request.Context(), term, params.Limit, params.Offset)
-
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(courses), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"term":       term,
-		"count":      len(courses),
-		"courses":    courses,
-		"pagination": pagination,
-	})
-}
-
-// GetCoursesByPrefix fetches courses by prefix within a term.
-func (h *Handler) GetCoursesByPrefix(c *gin.Context) {
-	term := normalizeTerm(c.Param("term"))
-	prefix := normalizePrefix(c.Param("prefix"))
-
-	if term == "" || prefix == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term and prefix parameters are required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	courses, hasNext, err := h.db.QueryByCoursePrefix(c.Request.Context(), term, prefix, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(courses), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"term":       term,
-		"prefix":     prefix,
-		"count":      len(courses),
-		"courses":    courses,
-		"pagination": pagination,
-	})
-}
-
-// GetCoursesByNumber fetches courses by prefix and number in a term.
-func (h *Handler) GetCoursesByNumber(c *gin.Context) {
-	term := normalizeTerm(c.Param("term"))
-	prefix := normalizePrefix(c.Param("prefix"))
-	number := normalizeCourseNumber(c.Param("number"))
-
-	if term == "" || prefix == "" || number == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term, prefix, and number parameters are required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	courses, hasNext, err := h.db.QueryByCourseNumber(c.Request.Context(), term, prefix, number, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(courses), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"term":       term,
-		"prefix":     prefix,
-		"number":     number,
-		"count":      len(courses),
-		"courses":    courses,
-		"pagination": pagination,
-	})
-}
-
-// GetCourseBySection fetches a specific course by term, prefix, number, and section.
-func (h *Handler) GetCourseBySection(c *gin.Context) {
-	term := normalizeTerm(c.Param("term"))
-	prefix := normalizePrefix(c.Param("prefix"))
-	number := normalizeCourseNumber(c.Param("number"))
-	section := strings.TrimSpace(c.Param("section"))
-
-	if term == "" || prefix == "" || number == "" || section == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term, prefix, number, and section parameters are required"})
-		return
-	}
-
-	course, err := h.db.GetCourseBySection(c.Request.Context(), term, prefix, number, section)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"course": course,
-	})
-}
-
-// SearchCourses runs a text search against courses for a term.
-func (h *Handler) SearchCourses(c *gin.Context) {
-	term := normalizeTerm(c.Param("term"))
-	query := strings.TrimSpace(c.Query("q"))
-
-	if term == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term parameter is required"})
-		return
-	}
-
-	if query == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Search query parameter 'q' is required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	courses, hasNext, err := h.db.SearchCourses(c.Request.Context(), term, query, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(courses), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"term":       term,
-		"query":      query,
-		"count":      len(courses),
-		"courses":    courses,
-		"pagination": pagination,
-	})
-}
-
-// GetTerms returns all known terms.
-func (h *Handler) GetTerms(c *gin.Context) {
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	terms, hasNext, err := h.db.QueryAllTerms(c.Request.Context(), params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(terms), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(terms),
-		"terms":      terms,
-		"pagination": pagination,
 	})
 }
 
@@ -295,250 +103,6 @@ func (h *Handler) GetAPIKey(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, apiKey)
-}
-
-// GetProfessorByID loads a professor by ID.
-func (h *Handler) GetProfessorByID(c *gin.Context) {
-	id := c.Param("id")
-
-	if id == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Professor ID is required"})
-		return
-	}
-
-	professor, err := h.db.GetProfessorById(c.Request.Context(), id)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get professor"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"professor": professor,
-	})
-}
-
-// GetProfessorsByName loads professors by name.
-func (h *Handler) GetProfessorsByName(c *gin.Context) {
-	name := c.Param("name")
-
-	if name == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Professor name is required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	professors, hasNext, err := h.db.GetProfessorsByName(c.Request.Context(), name, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get professors"})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(professors), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(professors),
-		"professors": professors,
-		"pagination": pagination,
-	})
-}
-
-// GetGradesByProfID loads grade distributions by professor ID.
-func (h *Handler) GetGradesByProfID(c *gin.Context) {
-	id := c.Param("id")
-
-	if id == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Professor ID is required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	grades, hasNext, err := h.db.GetGradesByProfId(c.Request.Context(), id, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get grades"})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(grades), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(grades),
-		"grades":     grades,
-		"pagination": pagination,
-	})
-}
-
-// GetGradesByProfName loads grade distributions by professor name.
-func (h *Handler) GetGradesByProfName(c *gin.Context) {
-	name := c.Param("name")
-
-	if name == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Professor name is required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	grades, hasNext, err := h.db.GetGradesByProfName(c.Request.Context(), name, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get grades"})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(grades), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(grades),
-		"grades":     grades,
-		"pagination": pagination,
-	})
-}
-
-// GetGradesByTerm loads grade distributions by term.
-func (h *Handler) GetGradesByTerm(c *gin.Context) {
-	term := c.Param("term")
-
-	if term == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term is required"})
-		return
-	}
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	grades, hasNext, err := h.db.GetGradesByTerm(c.Request.Context(), term, params.Limit, params.Offset)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get grades"})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(grades), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(grades),
-		"grades":     grades,
-		"pagination": pagination,
-	})
-}
-
-// GetGradesByPrefix loads grade distributions by prefix with optional term filter.
-func (h *Handler) GetGradesByPrefix(c *gin.Context) {
-	prefix := c.Param("prefix")
-
-	if prefix == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Prefix is required"})
-		return
-	}
-
-	term := c.Query("term")
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	var (
-		grades  []types.Grades
-		hasNext bool
-		err     error
-	)
-
-	if term == "" {
-		grades, hasNext, err = h.db.GetGradesByPrefix(c.Request.Context(), prefix, params.Limit, params.Offset)
-	} else {
-		grades, hasNext, err = h.db.GetGradesByPrefixAndTerm(c.Request.Context(), prefix, term, params.Limit, params.Offset)
-	}
-
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get grades"})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(grades), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(grades),
-		"grades":     grades,
-		"pagination": pagination,
-	})
-}
-
-// GetGradesByPrefixAndNumber loads grade distributions by prefix and course number with optional term filter.
-func (h *Handler) GetGradesByPrefixAndNumber(c *gin.Context) {
-	prefix := c.Param("prefix")
-	number := c.Param("number")
-
-	if prefix == "" || number == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Prefix and number are required"})
-		return
-	}
-
-	term := c.Query("term")
-
-	params, ok := parsePaginationOrRespond(c)
-	if !ok {
-		return
-	}
-
-	var (
-		grades  []types.Grades
-		hasNext bool
-		err     error
-	)
-
-	if term == "" {
-		grades, hasNext, err = h.db.GetGradesByPrefixAndNumber(c.Request.Context(), prefix, number, params.Limit, params.Offset)
-	} else {
-		grades, hasNext, err = h.db.GetGradesByNumberAndTerm(c.Request.Context(), term, prefix, number, params.Limit, params.Offset)
-	}
-
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get grades"})
-		return
-	}
-
-	pagination := buildPaginationMeta(params, len(grades), hasNext)
-
-	c.JSON(http.StatusOK, gin.H{
-		"count":      len(grades),
-		"grades":     grades,
-		"pagination": pagination,
-	})
-}
-
-// GetGradesBySection loads grade distributions for specific section
-func (h *Handler) GetGradesBySection(c *gin.Context) {
-	term := c.Param("term")
-	prefix := c.Param("prefix")
-	number := c.Param("number")
-	section := c.Param("section")
-
-	if term == "" || prefix == "" || number == "" || section == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Term, prefix, number, and section are required"})
-		return
-	}
-
-	grades, err := h.db.GetGradesBySection(c.Request.Context(), term, prefix, number, section)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get grades"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"grades": grades,
-	})
 }
 
 func normalizeTerm(value string) string {
